Created readlog.sh
This is the first version of readlog.sh. This file takes any new records from the current and previous hours PostgreSQL log files and shoves them into the logtail.csv. We then try to shove these into the PostgreSQL TimescaleDB database. If this works, then we delete the logtail.csv. If this fails, we continue to add more data to the logtail.csv until we can put the data into out PostgreSQL TimescaleDB database.